#105 refactor: 헤더, 푸터, 모달 스타일 디렉터리 경로 변경
