+++ frontend/src/components/Footer/FooterStyle.ts
